requirements.txt
alembic==1.14.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
certifi==2024.12.14
cffi==1.17.1
click==8.1.8
cryptography==44.0.0
greenlet==3.1.1
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
python-dotenv==1.0.1
rich==13.9.4
sentry-sdk==2.20.0
SQLAlchemy==2.0.37
typing_extensions==4.12.2
urllib3==2.3.0